一些编译警告及错误
git-svn-id: svn://23.27.117.66/h2/trunk@212 54bad739-14b1-1b4c-9595-0bb2987a2747

+++ lv/lv/StreamHelper.hpp
@@ -44,7 +44,7 @@ namespace lv
 				{
 					// std::showbase do not cooperate good with std::setw 
 
-					int width = os.width();
+					int width = static_cast<int>(os.width());
 
 					os << std::setw(0) << "0x" << std::setw(std::max(0, width - 2)) << std::noshowbase;
 				}
